+++ config/passport.js
@@ -34,7 +34,7 @@ passport.serializeUser((user, done) => {
 passport.deserializeUser(async (id, done) => {
     try {
         const user = await User.findById(id);
-        delete user.password;
+        delete user?.password;
         done(null, user);
     } catch (error) {
         done(error);