+++ controllers/author.js
@@ -1,8 +1,14 @@
 const Author = require('../models/Author');
 
-exports.authorLoggedInProcess = async (req, res) => {
+exports.getAuthorLoggedInProcess = async (req, res) => {
     if (!req.user) return res.status(200).json(null);
 
+    const { id } = req.user;
+    const user = await Author.findOne({ userId: id });
+    return res.status(200).json(user);
+};
+
+exports.getAuthorProjectsProcess = async (req, res) => {
     const { id } = req.user;
     const user = await Author.findOne({ userId: id }).populate({ path: 'projectsOwned' });
     return res.status(200).json(user);
@@ -13,7 +19,7 @@ exports.editProfileImageProcess = async (req, res) => {
     const { profileImage } = req.body;
     if (!profileImage) return res.status(400).json({ message: 'Send the image url' });
     const editedAuthor = await Author.findByIdAndUpdate(id, { profileImage }, { new: true });
-    return res.status(202).json({ message: 'Author updated', editedAuthor });
+    return res.status(202).json({ message: 'Author successfully updated', editedAuthor });
 };
 
 exports.editAuthorProcess = async (req, res) => {
@@ -42,5 +48,5 @@ exports.editAuthorProcess = async (req, res) => {
         propertiesToEdit[property] = req.body[property];
     });
     const editedAuthor = await Author.findByIdAndUpdate(id, { ...propertiesToEdit }, { new: true });
-    return res.status(202).json({ message: 'Author updated', editedAuthor });
+    return res.status(202).json({ message: 'Author successfully updated', editedAuthor });
 };