+++ models/Blob.js
@@ -1,0 +1,11 @@
+const mongoose = require('mongoose');
+const Schema = mongoose.Schema;
+
+const blobSchema = new Schema({
+    content: {
+        type: String,
+        required: true
+    }
+});
+
+module.exports = blobSchema;