+++ models/Project.js
@@ -1,5 +1,8 @@
 const mongoose = require('mongoose');
 const Schema = mongoose.Schema;
+const blobSchema = require('./Blob');
+const commitSchema = require('./Commit');
+const treeSchema = require('./Tree');
 
 const projectSchema = new Schema(
     {
@@ -7,10 +10,26 @@ const projectSchema = new Schema(
             type: Schema.Types.ObjectId,
             ref: 'Author'
         },
+        blobs: [blobSchema],
+        collaborators: [
+            {
+                type: Schema.Types.ObjectId,
+                ref: 'Author'
+            }
+        ],
+        commits: [commitSchema],
+        headVersion: String,
         privacyStatus: {
             type: String,
             enum: ['private', 'public'],
             default: 'private'
+        },
+        suggestions: [String],
+        trees: [treeSchema],
+        workStatus: {
+            type: String,
+            enum: ['composing', 'published'],
+            default: 'composing'
         }
     },
     {