+++ routes/auth.js
@@ -3,17 +3,19 @@ const router = express.Router();
 const { isAuth, catchErrs } = require('../middlewares');
 
 const {
-    loginProcess,
-    signupProcess,
-    // confirmSignupProcess,
     changePasswordProcess,
-    logoutProcess
+    // confirmSignupProcess,
+    deleteAccountProcess,
+    loginProcess,
+    logoutProcess,
+    signupProcess
 } = require('../controllers/auth');
 
-router.post('/login', loginProcess);
-router.post('/signup', signupProcess);
-// router.get('/confirm/:email/:id', catchErrs(confirmSignupProcess));
 router.post('/changePassword', isAuth, catchErrs(changePasswordProcess));
+// router.get('/confirm/:email/:id', catchErrs(confirmSignupProcess));
+router.delete('/delete', isAuth, catchErrs(deleteAccountProcess));
+router.post('/login', loginProcess);
 router.get('/logout', logoutProcess);
+router.post('/signup', signupProcess);
 
 module.exports = router;