+++ routes/author.js
@@ -2,9 +2,15 @@ const express = require('express');
 const router = express.Router();
 const { isAuth, catchErrs } = require('../middlewares');
 
-const { authorLoggedInProcess, editAuthorProcess, editProfileImageProcess } = require('../controllers/author');
+const {
+    editAuthorProcess,
+    editProfileImageProcess,
+    getAuthorLoggedInProcess,
+    getAuthorProjectsProcess
+} = require('../controllers/author');
 
-router.get('/', authorLoggedInProcess);
+router.get('/', getAuthorLoggedInProcess);
+router.get('/projects', isAuth, catchErrs(getAuthorProjectsProcess));
 router.patch('/profileImage/:id', isAuth, catchErrs(editProfileImageProcess));
 router.patch('/:id', isAuth, catchErrs(editAuthorProcess));