+++ changelog.md
@@ -2,3 +2,6 @@
 
 -   Feature: Basic project structure.
 -   Feature: Author and Project CRUD operations.
+-   Feature: Add Blob, Tree and Commit Schemas.
+-   Feature: Update CRUD operations.
+-   Feature: Add delete account